+++ tests/test_visualizations/test_feature_importance_plot.py
@@ -32,7 +32,7 @@ class TestFeatureImportancePlot:
     @pytest.fixture(scope="class")
     def regression_result(self) -> Result:
         """Setup a regression Result"""
-        dataset = load_demo_dataset("boston")
+        dataset = load_demo_dataset("california")
         model = Model(LinearRegression())
         return model.score_estimator(dataset)