+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - `model.test_estimators` with CV will keep using CV if it's refitting the best estimator.
 - `Result` now has a `.parameters` attribute to show what parameters generated the result
 - Switch to pyproject.toml for project metadata
+- Updated documentation to use california housing dataset instead of boston
+- Updated documentation to remove deprecated parameters from estimators
 
 # v0.12.0
 - Permutation importance and Feature importance are now two different plotting methods.